+++ src/main/java/com/woowacourse/calculator/Application.java
@@ -1,4 +1,4 @@
-package unit;
+package com.woowacourse.calculator;
 
 public class Application {
 	public static void main(String[] args) {
