+++ src/main/java/com/woowacourse/calculator/domain/Number.java
@@ -1,4 +1,4 @@
-package unit.domain;
+package com.woowacourse.calculator.domain;
 
 import java.util.Objects;
