+++ src/main/java/com/woowacourse/calculator/domain/Token.java
@@ -1,4 +1,4 @@
-package unit.domain;
+package com.woowacourse.calculator.domain;
 
 public interface Token {
 	public boolean equals(Object obj);
