+++ src/test/java/com/woowacourse/calculator/domain/OperatorTest.java
@@ -1,7 +1,6 @@
-package unit.domain;
+package com.woowacourse.calculator.domain;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
