+++ src/test/java/com/woowacourse/calculator/domain/TokenizerTest.java
@@ -1,4 +1,4 @@
-package unit.domain;
+package com.woowacourse.calculator.domain;
 
 import static org.assertj.core.api.Assertions.*;
